Allow float feature weights in item and user feature matrices.

+++ tests/test_fast_functions.py
@@ -8,7 +8,7 @@ from lightfm import lightfm_fast
 def test_in_positives():
 
     mat = sp.csr_matrix(np.array([[0, 1],
-                                  [1, 0]])).astype(np.int32)
+                                  [1, 0]])).astype(np.float32)
 
     assert not lightfm_fast.__test_in_positives(0, 0, lightfm_fast.CSRMatrix(mat))
     assert lightfm_fast.__test_in_positives(0, 1, lightfm_fast.CSRMatrix(mat))
